moved featureTransform into seperate file.
